+++ modules/common/Directories.qml
@@ -43,6 +43,8 @@ Singleton {
     property string shellConfig: FileUtils.trimFileProtocol(`${Directories.config}/illogical-impulse`)
     property string shellConfigName: "config.json"
     property string shellConfigPath: `${Directories.shellConfig}/${Directories.shellConfigName}`
+    property string updateLogPath: FileUtils.trimFileProtocol(`${Directories.state}/user/update.log`)
+    property string updateStatusPath: FileUtils.trimFileProtocol(`${Directories.state}/user/update-status`)
     property string todoPath: FileUtils.trimFileProtocol(`${Directories.state}/user/todo.json`)
     property string todoTxtPath: FileUtils.trimFileProtocol(`${Directories.state}/user/todo.txt`)
     property string notepadPath: FileUtils.trimFileProtocol(`${Directories.state}/user/notepad.txt`)